requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
pyjwt
pymysql
cryptography
flask-cors
